Create the masterService/driverRepository

+++ src/main/java/com/happyman/Ruby/masterService/repository/RoomRepository.java
@@ -2,6 +2,6 @@ package com.happyman.Ruby.masterService.repository;
 
 import com.happyman.Ruby.masterService.dao.Room;
 
-public abstract class RoomRepository implements BaseRepository<Room, Long> {
+public abstract class RoomRepository implements BaseRepository<Room, Integer> {
 
 }
